🚚 Move sanitize related code to utils for easier testing
