+++ utils/sanitize_test.go
@@ -1,4 +1,4 @@
-package db
+package utils
 
 import (
 	"bytes"
@@ -8,7 +8,7 @@ import (
 
 func TestSanitizeJSON(t *testing.T) {
 	input := []byte(`"\t\\\b\c\d\e\r\n\uD83D\uDe02\"\/\f\b\uDe02\uD83D\u0000\u0001XXXX\u012X"`)
-	output := sanitizeJSON(input)
+	output := SanitizeJSON(input)
 	expected := []byte(`"\t\\\bcde\r\n\uD83D\uDe02\"\/\f\b\u0001XXXX012X"`)
 	if !bytes.Equal(output, expected) {
 		t.Errorf("sanitizeJSON failed, expect %#v, got %#v", string(expected), string(output))
@@ -442,7 +442,7 @@ func TestSanitizeJSON(t *testing.T) {
   }
 }
 `
-	output2 := sanitizeJSON([]byte(input2))
+	output2 := SanitizeJSON([]byte(input2))
 	expected2 := []byte(strings.ReplaceAll(input2, `\ud83d`, ``))
 	if !bytes.Equal(output2, expected2) {
 		t.Errorf("sanitizeJSON failed, expect %#v, got %#v", string(expected), string(output))
